Another attempt at solving the problem I'm having with opening the folder dropdown in Travis CI during acceptance testing.
Perhaps the problem is that Capybara is, in some tests, trying to open the dropdown without giving it time to close from a previous action. Added an assertion that the menu is closed before clicking on the button to open it, to ensure that capybara waits for the menu to close before opening it again. I'm a bit in the dark here...