Immediately change background color of entries when user changes their state (read/unread) and then change it again when the server confirms that the state change has been completed successfully.
The difference between colors is subtle so that it doesn't annoy the user. E.g. when a user marks an entry as unread it turns a grey-and-slightly-greenish color immediately, an Ajax POST is sent to request the state change, and when the server responds with an HTTP 200 (and a JSON document), the entry turns grey-with-no-green. Both colors are very similar and the user probably won't even notice.

This is achieved by immediately applying a CSS class as soon as the user performs the action on the page, and a different CSS class when the response from the server is received.

This change is necessary because of the change in commit b08bcd9419319ea949c09c697a2a03bee6dac07e. For acceptance tests to pass, we need to know when the server has finished changing the entry state, so we have to apply a different CSS class when the server response is received. However it is also necessary to change what the user is seeing as soon as he clicks on the page to give a good responsivity feeling; otherwise the server response could take several seconds to arrive and user perception would be that the application is slow and unresponsive.

The compromise reached is: apply a CSS class as soon as the user clicks on the page (so that he's happy with responsivity) and apply a different one as soon as the AJAX response is received (enabling acceptance tests to know when the entry has actually changed state). Also, both CSS classes are similar enough not to be noticeable to the user in most cases.

I could have made the background-color in both CSS classes the same, but I think in some cases it could be useful to know when the server response has arrived for a user that is aware of this feature.

Actually two new CSS classes have been created: one for when an entry is in the process of being marked as read, and another one for when it's in the process of being marked as unread.