Removed the before_install step from the Travis CI configuration.
It should no longer be necessary to manually install the latest bundler gem every time a build is performed; this was necessary for some time because of problems in Travis infraestructure that should now be solved.

Also it's not necessary to set the DISPLAY shell environment variable, nor running xvfb to have a virtual display buffer. Capybara-webkit, which is our current javascript driver for acceptance tests, is fully headless and shouldn't need a display buffer.

This should speed up CI builds a bit.