Added back Resque config file, after having deleted it from Git history.
The re-added version has no information about Staging and Production environments. This information is stored in different versions of this file which are stored in the servers themselves, never in git, and which are linked during deployment.

This way the Redis server address/port in each environment is not readable from git history.