Removed resque-web watch from God configuration.
It is not necessary that God launches and monitors a standalone resque-web instance in the background server. This resque-web instance would never be accessed. Instead a resque-web instance is automatically launched in the frontend server just by mounting it in routes.rb under the /admin/resque path.

The resque-web instance in the frontend server works correctly and displays the right data because it accesses the redis server in the background server, where it can read all relevant information about resque state in that server.