Created new FeedSubscription model.
We're going to change the association between feeds and users. Instead of a simple has-and-belongs-to-many association (which implies a join table in the database to which no additional columns can be added) we will have a more complex has-many-through association. This implies creating a new FeedSubscription model, which will join users with feeds. This model is backed by a table similar to the old join table, but we can add additional columns to the table which will be available as attributes of the model.

This is necessary because we want to store, for each user-feed subscription, how many unread entries the user has in the feed. This will greatly speed up the application: instead of having to count unread entries each time the user loads the page, the current count of unread entries will be already saved in the database.

This count of unread entries will be updated in background every time an entry changes state (read/unread), instead of having to be computed every time the user loads the page.

For the moment only the FeedSubscription model, along with its unit test and database migration to create its table, have been implemented. User and Feed models have not been touched yet.

Next step will be to change the User and Feed models to redefine their association to be through the FeedSubscription model; a database migration will be implemented to remove the feeds_users table, with will be no longer necessary. This will probably mean that code like this:

    user << feed

will probably stop working, something like this will have to be done instead:

    feed_subscription = FeedSubscription.new user_id: user.id, feed_id: feed.id
    feed_subscription.save!

This means that a lot of things will be probably broken by the change and will have to be fixed. Oh, the joys of refactoring!