WIP: proper management of the cache of the count of unread entries.
Minor renaming of some classes, methods and parameters.

Logic to add a new feed subscription to a user has been moved to the SubscriptionsManager class.

Logic to remove a subscription has been moved to the SubscriptionsManager class. FeedUnsubscriber class has been deleted because it's now redundant.

When trying to find out if an entry is unread for a user, and the user is not subscribed to its feed, an error is raised. I think it's better to raise an error in this case instead of returning a default value, which would mask a logic error.