WIP: unit tests for the EntryState model, to check its lifecycle callbacks.
The idea is that most changes in the unread_entries attribute of the FeedSubscription model will be triggered by changes in EntryState model instances (creating, deleting, or changing the "read" attribute of such instances). The only exception will be the creation and deletion of the FeedSubscription database record themselves, which will be triggered by the act of a user subscribing/unsubscribing from a feed.