The unread entries count besides each feed in the sidebar is obtained from the pre-calculated field in the feed_subscriptions table, instead of being calculated on each page load.
Also, a chech has been added to ensure that the user is actually subscribed to the feed before painting it in the sidebar. This is to solve a possible data inconsistency in which a user is not subscribed to a feed, but the feed is in one of his folders. This should never happen on normal use but could be caused by direct database manipulation by an admin. Of course this additional check slows down a bit loading the app, so it may be removed in the future if it seems wise to do so.