Instead of reloading user model in some controller actions (introduced in commit f837a55505cb8e05ea504a6978e2a793e90bc650), the user model is always reloaded in SubscriptionsManager#folder_unread_count.
This simplifies code: instead of having to reload the model instance in many places (just before invoking SubscriptionsManager#folder_unread_count), it is reloaded in a single place (inside the method). We always want to reload it in case the unread_entries column has changed in the database, anyway.

For the same reasons, the folder instance is also reloaded inside SubscriptionsManager#folder_unread_count (in case an actual Folder instance is passed to the method, instead of the magic "all" string).