Added new secret_key_base config parameter to initializers.
This is a new Rails4 config option that, if present, encripts the client-side cookie store to make it unreadable as well as tamper-proof.

It has been changed in the initializer file used for Travis CI, which is present in git, but also in the initializer files used in real environments, which are not in git.