Removed log4r gem because of incompatibilities with Rails 4.
This means going back to the default Rails logging.

The incompatibility is caused because Rails added a new "formatter" method that logging frameworks must implement. Log4r does not implement it; when trying to start a Rails 4 app using log4r logging, an "undefined method 'formatter'" error is raised and the startup fails. Highest criticality!

It seems logging frameworks (yell, log4r) have not adapted yet despite Rails 4 being released a month and half ago (and RCs long before that). There's a pull request in the log4r github repo supposedly fixing this that has received no attention for a month. I suppose people are happy enough with Rails logging, because obviously logging frameworks are not getting enough love. I'm really sorry, log4r is very nice, but I just can't hold migrating to Rails 4 because of this, and I've lost confidence in log4r devs adapting to the state of the art. Back to the default logger it is.