Added strong parameters security to FeedsController and FoldersController.
The structure of the JSON sent via AJAX to these controllers has been simplified a bit. Only really needed parameters are sent, instead of serializing the whole form to POST it. Also a hidden field in a form has been removed, its value can be included more easily in the AJAX POST with coffeescript code.