Use PATCH HTTP verb to request a feed refresh.
Also instead of a custom named route for /feed/:id/refresh which maps to FeedsController#refresh, the more usual resourceful style is used, with an :update route for the feed resource mapped to FeedsController#update.

Before this commit we were doing feed refreshes with a GET, which is not a good idea as GET should be idempotent; refreshing a feed clearly isn't. PATCH is more appropiate to an action which can potentially change data in the server. Rails 4 support for PATCH enables us to make better use of HTTP verbs.

For the moment in Rials 4.0 all actions done with PATCH are also done with PUT; we use PATCH instead of PUT in all cases because this PUT support may be deprecated in the future, read:

  http://weblog.rubyonrails.org/2012/2/25/edge-rails-patch-is-the-new-primary-http-method-for-updates/

As a side effect, the path to read a feed and to request a refresh is the same; the difference is that the URL is invoked with a GET or with a PATCH. This enables us to remove a global javascript variable (which previously stored the refresh URL); now we only keep a global variable with the URL for the curretnly selected feed. This URL can be used both for retrieving the feed and for requesting its refresh.