Moving a feed into a folder is done with a single method, FolderManager.move_feed_to_folder, regardless of whether the folder already exists or not.
For this, a facade method has been created, FolderManager.move_feed_to_folder, and the previously existing methods have been made private. The facade method receives optional arguments that tell it if the folder already exists or not, and then it calls the correct private method.

From the point of view of classes using the services of FolderManager now things are a bit different; instead of two methods, there's a single method with optional arguments.

The optional arguments are named arguments, and therefore do not work in versions of Ruby older than 2.0