WIP: remove FoldersController#remove action.
Instead a feed will be removed from a folder by calling the FoldersController#update action with the param folder_id to the value "none".

Already implemented: the same method is invoked on the User instance to move a feed to another folder and to remove it from its current folder. FoldersController#remove and FoldersController#update actions are exactly the same and could be simplified keeping only one of them.

Not yet implemented: javascript code should call FoldersController#update with a special value of "none" in the folder_id param.