Removing a feed from its folder is requested via an HTTP PATCH request, instead of a DELETE request.
The path for this PATCH request is the same as when moving a feed to an existing folder, and it's routed to the FoldersController#update RESTful action. The difference is that when removing a feed from a folder the folder id passed as a parameter is "none".

The custom named route "folder_remove" has been deleted along with the FoldersController#remove action. Only normal RESTful routes and controller actions are used now for folders.