Global GUID uniqueness for entries is no longer enforced. GUIDs are in the scope of a single feed, but entries from different feeds can have the same GUID.
This means that an entry is updated only if an entry is retrieved with the same GUID when fetching the feed to which the entry belongs. If an entry with the same GUID is retrieved when fetching a different feed, it is treated as a different entry and saved independently in the database.