Use current_user instead of user_signed_in? to detect if the user is signed in.
Hopefully this helps with a bug I was experiencing: sometimes the root page was displayed as if the user was not logged in, but when clicking on "sign in" instead of being shown the login form the user was shown the feeds list, with an error message saying that "user was already logged in". This means the code in the root page that detects if the user is logged in was not working correctly.

There are several threads in stack overflow of people complaining about user_signed_in? not working, and some of them report that checking if current_user exists instead gives better results.