The home page is the same regardless of whether the user is logged in or not.
The difference is that if he's logged in the navbar will be shown, and the "sign in" link will take him to the feeds list instead of the login form.

This is desirable because this page was unable to detect that the user had a valid devise :rememberable cookie and therefore had to be treated as if he was logged in.