User instance methods now always receive model instances as arguments, instead of IDs.
It is now the responsability of the controllers to instance these models with the IDs passed as request parameters, and to ensure that users do not access models they shouldn't (e.g. a user can only move a feed to a folder if he's subscribed to the feed and owns the folder).

This change in paradign means similar changes in auxiliary classes in /lib too.