Simplified ImportSubscriptionsJob. Deleted FetchImportedFeedJob.
Instead of trying to do the work of detecting if each imported feed already exists, enqueuing a job to fetch it if it doesn't, etc in ImportSubscriptionsJob, the User#subscribe method is used. This method accepts a URL and does all of that and more, ensuring sanity of data and other bells and whistles. The User#subscribe method fetches the feed if necessary, which makes FetchImportedFeedJob totally unnecessary.

As a side effect, the data import process is totally sequential: each feed is processed and fully fetched if necessary, and until it's not been fully processed the job does not continue with the next feed. This should not be a problem; users should expect the import process to be lengthy anyway.