When saving an entry, if the url attribute is not a valid URL but the guid attribute is, the url attribute takes the value of the guid attribute.
This is intended for feeds like this one:

  https://www.tbray.org/ongoing/ongoing.atom

in which the <link> for each entry is just a relative URL but the <id> is a valid absolute URL which we can use directly.

From what I've seen, using the entry's absolute URL as its ID/GUID is quite common (though not always done).