Rolled back to the official Feedzirra 0.2.0.rc2 version.
This implies a rollback to sax-machine 0.2.0.rc1 official and Nokogiri 1.5.2

There is a problem when using latest Feedzirra from my repo with Nokogiri 1.6.0 which caused some feeds to be incorrectly parsed. For instance the feed from meneame.net is interpreted by Feedzirra as if it were an itunes feed, which it isn't.

This is probably an incompatibility of this old code with nokogiri 1.6.0. I should fix this when I fork Feedzirra to remove its dependency from sax-machine and make it directly dependant on Nokogiri (last version).