Simplified Simple_form initialization.
We only use one simple_form wrapper in all the app, which is supposed to generate markup compatible with bootstrap forms with prepended icons (even though it's not currently working because of the migration to Bootstrap 3). Removed from the app initializers all other wrapper definitions, which were unused clutter. Also set the only wrapper as the default one, so that each form does not have to explicitly declare which wrapper to use.