Simple_form adds class "form-horizontal" to forms by default.
This removes the need to explicitly declare it in each form.

Also the "btn btn-default" classes are used by default for buttons, instead of just "btn". This is necessary because buttons in Bootstrap 3 no longer have styling unless given the "btn-default" or alternative class.

Also enabled browser-side validations by default.