Use the github repo as source for Simple_form gem.
We need support for the config.input_class option in simple_form. This has been added in the last 15 days and is not yet present in any tagged version. As soon as a tagged version or, even better, a stable release has this support, we will switch back to using a released gem or tag.

This is necessary to add "form-control" class to all inputs in forms, which is a requisite for Bootstrap 3 compatibility.

For more details about integration between simple_form and bootstrap 3 see:

    http://stabco.tumblr.com/post/59760641051/simple-form-bootstrap3-integration