Rollback to using simple_form helper to create the OPML upload form.
This way the authenticity token is automatically put in a hidden input that is POSTed with the form.

The markup of the file input, which is the only visible input in the form, is created by hand instead of with form helpers. The reason is that this input is the only one that looks ugly if the Bootstrap 3 "form-input" class is applied to it; the current simple_form wrapper I've written applies this class to all inputs; and it's not worth it defining a custom input because this is the only file input in the whole app. It's easier to just write the markup by hand in this case.