Fixed OPML upload form.
Use simple_form helper to generate the file input markup. It seems creating the markup by hand does not work properly, the file name gets POSTed instead of the file itself. I could probably make it work, but the helper is functionally complete so it's not worth investigating further.

Added CSS rules so that the file input is displayed properly instead of trying to add a border like other bootstrap inputs.