Changed the sidebar markup to fix the refresh feed functionality.
This was broken on the migration to Bootstrap 3. The reason is that the javascript handler launched when clicking on a feed was searching for a data-* attribute in the wrong place. As a result, the currently open folder was not correctly stored in a global variable. When clicking on the "refresh feed" button, the currently open folder was not known and therefore the app was stuck in the "loading.." message forever.

The fix is just to move the data-* attribute to a place where the javascript handler function finds it. This attribute was not used anywhere else, so no other changes are necessary.