Fixed code that gives autofocus to the first text input field in all popups as soon as they are shown.
With the migration from Bootstrap 2 to 3, the event name has changed (before it was "shown", not it's "shown.bs.modal"). Coffeescript event handlers have been changed accordingly.

Also added code to give focus to the cancel button in those popups that do not have text inputs. The reason is that this way the popup can be closed with the ESC key. It seems that unless a popup input has the focus, the ESC key does nothing.