Fix acceptance tests.
Removed acceptance test for the removed "my feeds" link in the navbar.

Made the dropdown with the "my account", and "logout" links option so that it only appears if a user is logged in. It's a bit redundant, because the navbar itself should not appear unless a user is logged in, but it doesn't hurt to program defensively.