Moved the main application view to its own /read path, with its own controller.
The idea is that soon RESTful resources will only respond to JSON requests with raw data; it will be the responsability of the main application page (or other clients in the future) to render that data and include it in the view HTML. It is envisioned that the main app page will use angularJS for this.

Therefore it makes sense to separate json views, associated with proper RESTful resources, from the main application view, which serves a bunch of dynamic HTML which through AJAX retrieves JSON data about the resources.