The divs that contain alerts and sidebar folders are not displayed until angularJS has finished bootstrapping.
This is achieved by adding the ng-cloak attribute to the divs. Also we have to use a CSS rule that hides elements marked as ng-cloak; we wouldn't need it if the angularJS file was loaded in the page header, but it is loaded at the end of the page for performance reasons and therefore we must include such CSS rule in our SCSS files.

For more about ng-cloak directive see:

    http://docs.angularjs.org/api/ng.directive:ngCloak