When the start link in the sidebar is clicked, the global currently selected feed is set to null.
This is part of the transition to angularJS: instead of registering event handlers to change the DOM (triggering classes or whatnot), ng- attributes are added that call methods in the controller, which set model values in the scope. Then visual properties in the page are made dependant on those model values.

This way the DOM is not directly changed by event handler. Rather event handlers change values in the model, and that triggers changes in the DOM automatically.