The FeedsController#destroy action returns just an HTTP 200 if successful.
After the migration to angularJS, there is no need for the server to return JSON with the unread entries of folders after unsubscribing, whether the folder in which was the unsubscribed feed has been deleted, etc. All this data is known by the angularJS code running in the client, which can update the view without receiving this information from the server. All the client needs is a success/failure response from the server to know whether to display an error message or not.