Removed alert that appeared when user tried to move a feed to the folder it's already at.
This alert was supposed to appear when the server responds with an HTTP 304 when moving a feed to a folder, but this never actually happens anymore. If the server is requested to move a feed from a folder to the same folder, nothing actually happens at the database level but the server responds happily with an HTTP 200. Therefore all the code related to showing an alert when a 304 is received can be safely removed.

This is actually desirable. It makes no sense to tell the user an error has happened simply because he's clicked on an already checked folder in the dropdown. It's better to silently do nothing.