Fixed devise alerts to work with angularjs.
Devise alerts are closed automatically after 5 seconds using code in the angularjs controller.

For this to work, all views (including devise and the page served by High_Voltage) are angularjs apps.

Also, fixed the application.js manifest so that the ng-app.js.coffee file is loaded before any others in the app/assets/javascript/modules folder. This is necessary because if another angularjs file is loaded before ng-app (for example the controller files), an error will happen when they try to create controllers etc in a module which has not yet beed defined (the "feedbunch" module is defined in ng-app). Therefore ng-app must be the first javascript file loaded by the assets pipeline.