Marked an acceptance test as pending for the moment.
We're going to change the logic behind unread entries counts. This count will only be kept for feeds, and the counts for users and folders will be removed. The numbers shown in the view for folders and for the logged in user will be calculated by AngularJS based on the unread entries counts of feeds.

For the moment this does not work, and this is why this test (which tests among other things the unread entries counts in folders after refreshing a feed) currently fails.