Reverted changes in commit a8d17139aca257b731e0c551033dac65d7d6d38a.
Restored a functionality from before migrating to angularjs: while an entry state is changing, it is marked with the CSS class "entry-becoming-read" or "entry-becoming-unread". Only when an HTTP OK response has been received from the server the class becomes "entry-read" or "entry-unread".

This enables acceptance tests to look at the DOM and know when the state change operation has actually finished.