User#refresh_feed no longer returns the list of entries. It will raise a NotSubscribedError if it is passed a feed to which the user is not subscribed.
This means application logic must change; when requesting a feed refresh, after receiving the "OK" response the client will have to request its entries, the same as if the user clicked on a feed to read its entries. This simplifies pagination: pagination logic can be onl associated with entries retrieval, not with feed refreshing.