When clicking on a feed or folder, smoothly scroll to top.
This is useful when having a very long list of entries and clicking on one at the bottom with few entries. In this case the entries list would load above the viewport and the user could be confused. Thanks to this change the page will autoscroll to the top and he will see the entries.

In some cases I guess it could be annoying, but right now I think usability advantages outweight problems. I'll keep an eye on it, this behavior should be considered "on trial".