When clicking on a feed or folder, subscribing to a feed or unsubscribing from a feed, the page automatically scrolls to the top.
This is necessary in most cases to surprise the user as little as possible. E.g. when adding a subscription or clicking on a feed he expects to see the entries list from the newest one onwards, and when he unsubscribes he expects to see the start page from the top.

In some cases it could be annoying to some users, but I think this behavior is mostly good.

The scrolling is done with jQuery code, put inside an angularjs service. A bit messy perhaps but it works and is extremely simple; any other solution would require more code.