Remove the loading_entries boolean flag, which indicated that the first page of entries was being loaded.
This is not necessary and it is redundant with the loading_entries_busy flag, which indicates that an entries page is being loaded. If we need to check if the first page is being loaded, it is enough to check the loading_entries_busy flag (which would indicate that entries are being loaded) and the entries_page root scope variable, which would tell us if the page being loaded is actually the first one.

Therefore the loading_entries flag is totally redundant and has been removed.