When alerts in the read view are closed, they can be displayed again if necessary (e.g. the user clicks on a feed without entries several times).
When an alert is displayed any running timers for that alert are reset. This means that e.g. if the user clicks several times on a feed without entries, the alert will be displayed for 5 seconds after the last click.

When an alert is manually closed, any running timers for that alert are cancelled. This means that e.g. if the user clicks on a feed without entries, closes the alert and then quickly clicks on the feed again, the alert will be displayed for 5 seconds after the second click, instead of 5 seconds counting from the first click.

All this behavior only affects alerts that only appear in the read view. Other alerts that can appear in any view, caused by rails or devise errors, are not susceptible of being displayed again without a page reload, and therefore do not need this behavior. They are simply closed using Bootstrap data-* attributes.