Removed "you must sign in" alert message.
The way authentication is set up, this message appeared every time a user tried to sign in when no error had in fact happened.

If the user tries to access a protected page while unauthenticated, he will be redirected to the login page. This should be enough for him to get the hint, there's no need for an additional alert.