Refined the code that automatically opens/closes folders in the sidebar when moving feeds around.
It's still a little messy though. The Bootstrap "collapsible" plugin was clashing with angularjs code. Also the accordion-like behavior got broken after programatically collapsing folders; according to the documentation this should be avoided by passing an options object with the correct "parent" key, but I haven't been able to make it work.

Anyway, it seems that manually toggling the "in" and "collapse" classes gets the job done. There is no open/close animation (change is instantaneous) but it does what is intended and the accordion-like behavior is not broken. Also I've had to introduce a small delay to give time for angularjs to "stabilize" after model changes.

I know, it's messy. I might still remove all this behavior if it gives me more trouble. It's even a bit uglier than I'd like for the end user, but hopefully moving feeds around, unsubscribing etc are actions that are not performed often, so the user won't have to see this hack moving but rarely.

Don't judge me.