Unread entries badges display "999+" if there are more than 999 unread entries.
This is achieved with a new angularjs filter.

This is desirable to keep badges from getting arbitrarily wide. If badges get too wide, at a certain poing they "jump" to a second row in the sidebar, making it look ugly. By limiting their width it can be ensured that the sidebar is properly displayed in all cases.