Changed the text size limit for the "content" column in the "entries" table to 16777215. This should tell MySQL that a "mediumtext" data type is to be used instead of "text".
This is intended to solve the problem with some entries that were too long for a text column, causing an error when fetched. Mediumtext columns have a 16 megabyte size limit instead of the 64kb of text columns. I don't think entries longer than 16Mb are reasonable, mediumtext should be big enough (there's also the option of longtext).