More robust detection of when the import process has finished.
The total_feeds and processed_feeds columns are still used; if the number of processed feeds is equal to the total number of feeds in the OPML, we know that
the import process has finished.

However SubscribeUserJob now also checks:

- if there are other instances of SubscribeUserJob still enqueued (which happens if there are other jobs from the import waiting to run)
- if there are other instances of SubscribeUserJob or any instance of ImportSubscriptionsJob currently running, which could happen if there are multiple workers running the import jobs.

This new code should not normally be triggered, and the detection on whether the import is finished is normally based on the total_feeds and processed_feeds columns. However if some of the jobs are killed (e.g. with an out of memory condition) without being able to update the database, the new code should detect when the import has finished anyway.