Locale parameter is not automatically added to the href of all links anymore.
The locale param is manually added to those links where it's necessary (mostly devise views).

The idea is that while the user is not authenticated, the locale param is added to links, so that if the user changes the locale from the dropdown the change will be preserved if he changes to another devise un-authenticated view (change password, send unlock email etc). However as soon as he's authenticated and passes to the read view, the locale used is the one set for the user, and the param is not necessary in links (in fact it could be harmful).