The edit registration view uses the "application" layout, instead of the "devise" one.
The edit registration view looked weird with both the navbar and the big image header used in other devise views. The devise layout is appropiate for views that are accessed by unauthenticated users; the edit registration view, which is accessed by authenticated users, is more suited to the application layout.

Inspired by this article of the devise wiki:

  https://github.com/plataformatec/devise/wiki/How-To%3a-Create-custom-layouts